Makes Simulations directory, moves TimeStepWizard to Simulations
