+++ src/Utils/Utils.jl
@@ -18,7 +18,6 @@ include("units.jl")
 include("loop_macros.jl")
 include("launch_config.jl")
 include("cell_advection_timescale.jl")
-include("time_step_wizard.jl")
 include("pretty_time.jl")
 include("pretty_filesize.jl")
 include("tuple_utils.jl")
